EC/backend/API/Image_to_Text/requirements.txt
fastapi==0.109.2
uvicorn[standard]
transformers==4.37.2
torch==2.2.0
pillow==10.2.0
streamlit==1.31.1
ui==0.1.4

transformers==4.23.2
requests==2.31.0
Pillow==9.4.0]
Flask==3.0.2
Flask-Cors==4.0.0
torchaudio==2.2.0
numpy==1.26.4
pandas==2.2.0